feat: update .env.example and seed module to use @nestjs/config

+++ src/seed/seed.service.ts
@@ -1,13 +1,22 @@
-import { Injectable, Logger } from '@nestjs/common'
+import { Injectable, Logger, UnauthorizedException } from '@nestjs/common'
+import { ConfigService } from '@nestjs/config'
 
 const logger = new Logger('SeedService')
 
 @Injectable()
 export class SeedService {
-  constructor() {}
+  private isProd: boolean
+
+  constructor(private readonly configService: ConfigService) {
+    this.isProd = configService.get('STATE') === 'prod'
+  }
 
   async executeSeed(): Promise<boolean> {
+    if (this.isProd) {
+      throw new UnauthorizedException('We cannot run SEED on production.')
+    }
     // * Limpiar la base de datos BORRANDO TODO
+
     await this.deleteAll()
     // * Crear los datos de ejemplo
     await this.createData()
